hid sensitive data to config/added relationship between users and posts

+++ app/routers/post.py
@@ -8,8 +8,8 @@ router = APIRouter(prefix="/posts", tags=["posts"])
 
 
 @router.get("/", response_model=List[schemas.PostResponse])
-def get_posts(db: Session = Depends(get_db)):
-    posts = db.query(models.Post).all()
+def get_posts(db: Session = Depends(get_db), limit: int = 10):
+    posts = db.query(models.Post).limit(limit).all()
     return posts
 
 
@@ -25,8 +25,7 @@ def get_one_post(id: int, db: Session = Depends(get_db), current_user: int = Dep
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=schemas.PostResponse)
 def create_post(post: schemas.PostCreate, db: Session = Depends(get_db),
                 current_user: int = Depends(oauth2.get_current_user)):
-    print(current_user.email)
-    new_post = models.Post(**post.model_dump())
+    new_post = models.Post(owner_id=current_user.id, **post.model_dump())
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
@@ -41,6 +40,9 @@ def delete_post(id: int, db: Session = Depends(get_db), current_user: int = Depe
     post = post_query.first()
     if post == None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"post with id {id} does not exist")
+
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="NOT AUTHORIZED to perform such actions")
 
     post_query.delete(synchronize_session=False)
     db.commit()
@@ -57,6 +59,8 @@ def update_post(id: int, updated_post: schemas.PostCreate, db: Session = Depends
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"post with id {id} doesn't exist")
 
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=f"NOT AUTHORIZED to perform such actions")
     post_query.update(updated_post.dict(), synchronize_session=False)
     db.commit()
